Refactor code
Create a custom hook, move two functions outside the main component
