+++ src/pages/Home.jsx
@@ -3,27 +3,14 @@ import "../styles/css/Home.css";
 import Card from '../components/Card';
 import Filters from '../components/Filters';
 import {useNavigate} from "react-router-dom";
-import * as dayjs from "dayjs";
-const customParseFormat = require('dayjs/plugin/customParseFormat');
-dayjs.extend(customParseFormat);
+import { useRequests } from '../functions/useRequests';
+import { filtrate } from '../functions/filtrate';
+import getHash from '../functions/getHash';
 
 
 export default function Home({showFilters}) {
-    // eslint-disable-next-line no-sequences
-    const getQueryParams = () => window.location.search
-        .replace('?', '')
-        .split('&')
-        .map(item => {
-            const data = item.split('=');
-            if (data.length && data[0] === 'hash') {
-                return {hash: data[1]}
-            }
-            return {};
-        });
 
-    const queryData = getQueryParams();
-
-    const hash = queryData.length && Object.keys(queryData[0]).includes('hash') ? queryData[0]['hash'] : null;
+    const hash = getHash();
 
     const navigate = useNavigate();
 
@@ -34,37 +21,8 @@ export default function Home({showFilters}) {
         }
     }, [hash, navigate])
 
-    const [requests, setRequests] = useState([]);
+    const { requests } = useRequests();
 
-    useEffect(() => {
-        fetch(process.env.REACT_APP_API_PREFIX + "/api/requests")
-            .then(response => response.json())
-            .then(data => {
-                console.log(data);
-                data = formatDate(data);
-                setRequests(data);
-            })
-            .catch(error => {
-                console.error(error);
-            })
-    }, []);
-
-    useEffect(() => {
-        setCards(requests);
-    }, [requests])
-
-    function formatDate(arr) {
-        arr.forEach(obj => {
-            obj.dateFrom = dayjs(obj.dateFrom, 'DD-MM-YYYY', true).toDate();
-            obj.dateTo = dayjs(obj.dateTo, 'DD-MM-YYYY', true).toDate();           
-        })
-        return arr;
-    }
-
-    //array of cards that will be displayed after filtration
-    const [cards, setCards] = useState([...requests])
-
-    
     const [filtrationParams, setFiltrationParams] = useState(
         {
             'from': null,
@@ -74,18 +32,11 @@ export default function Home({showFilters}) {
             "isRewardable": null
         });
 
+    const [requestsFiltered, setRequestsFiltered] = useState([...requests])
+
     useEffect(() => {
-        let allArray = requests;
-        let filter = filtrationParams;
-        let newArray = allArray.filter(function (el) {
-            return ((el.from === filter.from) || filter.from === null)  &&
-                   ((el.to === filter.to) || filter.to === null) &&
-                   (filter.dateFrom === null || (el.dateFrom.getTime() >= filter.dateFrom.getTime())) &&
-                   (filter.dateTo === null || (el.dateTo.getTime() <= filter.dateTo.getTime()));
-          });
-        setCards(newArray)        
-    }, [filtrationParams, requests])
-    
+        setRequestsFiltered(filtrate(requests, filtrationParams))
+    }, [filtrationParams, requests])  
 
   return (
     <div className='home-main'>
@@ -94,11 +45,11 @@ export default function Home({showFilters}) {
            
             {showFilters ? '' : 
                 <div className='cards-container'>
-                    {cards.length === 0 ? 
+                    {requestsFiltered.length === 0 ? 
                         <div><span>Unfortunately, there are no results for your query. 
                                 Try changing the filters
                         </span></div> 
-                     : cards.map((request, key) => (
+                     : requestsFiltered.map((request, key) => (
                         <Card key={key} {...request}/>
                     ))}
                 </div>}
