+++ src/styles/css/Filters.css
@@ -81,6 +81,25 @@ select option {
   color: #4A5567 !important;
 }
 
+.react-datepicker__close-icon::after {
+  background: rgb(45, 212, 191) !important;
+  display: flex !important;
+  align-items: center !important;
+  justify-content: center !important;
+}
+
+.react-datepicker__close-icon {
+  position: static !important;
+}
+
+.react-datepicker__input-container {
+  display: flex !important;
+  align-items: center;
+  gap: 5px;
+  flex-direction: row-reverse !important;
+  transform: translateX(-30px);
+}
+
 @media (max-width: 480px) {
   .filters-main {
     display: none;
